Move CA-Reduction to compressed-air
