+++ src/app/calculator/compressed-air/compressed-air-list/compressed-air-list.component.html
@@ -4,6 +4,21 @@
         <div class="d-flex calc-item-row w-100">
             <div class="p-2 col-2 d-flex justify-content-center align-items-center">
                 <div class="calc-icon-div d-flex justify-content-center align-items-center">
+                    <img src="assets/images/calculator-icons/utilities-icons/compressed-air-reduction-icon.png"
+                        class="calc-img" [routerLink]="['/calculators/compressed-air-reduction']">
+                </div>
+            </div>
+            <div class="d-flex flex-fill flex-column justify-content-center">
+                <a [routerLink]="['/calculators/compressed-air-reduction']">Compressed Air Reduction</a>
+                <p>
+                    This calculator is used to quantify the energy savings associated with reducing compressed air
+                    usage.
+                </p>
+            </div>
+        </div>
+        <div class="d-flex calc-item-row w-100">
+            <div class="p-2 col-2 d-flex justify-content-center align-items-center">
+                <div class="calc-icon-div d-flex justify-content-center align-items-center">
                     <img src="assets/images/calculator-icons/compressed-air-icons/leakage-loss-estimator.png"
                         class="calc-img" [routerLink]="['/calculators/bag-method']">
                 </div>
