+++ src/app/calculator/compressed-air/compressed-air-reduction/compressed-air-reduction-results/compressed-air-reduction-results.component.css
