+++ src/app/calculator/compressed-air/compressed-air-reduction/compressed-air-reduction.component.css
