+++ src/app/calculator/compressed-air/compressed-air.module.ts
@@ -10,6 +10,7 @@ import { ReceiverTankModule } from './receiver-tank/receiver-tank.module';
 import { SystemCapacityModule } from './system-capacity/system-capacity.module';
 import { CompressedAirListComponent } from './compressed-air-list/compressed-air-list.component';
 import { RouterModule } from '@angular/router';
+import { CompressedAirReductionModule } from './compressed-air-reduction/compressed-air-reduction.module';
 
 @NgModule({
   imports: [
@@ -22,6 +23,7 @@ import { RouterModule } from '@angular/router';
     PneumaticAirModule,
     ReceiverTankModule,
     SystemCapacityModule,
+    CompressedAirReductionModule,
     RouterModule
   ],
   declarations: [
