+++ src/app/calculator/utilities/utilities-list/utilities-list.component.html
@@ -116,21 +116,6 @@
         <div class="d-flex calc-item-row w-100">
             <div class="p-2 col-2 d-flex justify-content-center align-items-center">
                 <div class="calc-icon-div d-flex justify-content-center align-items-center">
-                    <img src="assets/images/calculator-icons/utilities-icons/compressed-air-reduction-icon.png"
-                        class="calc-img" [routerLink]="['/calculators/compressed-air-reduction']">
-                </div>
-            </div>
-            <div class="d-flex flex-fill flex-column justify-content-center">
-                <a [routerLink]="['/calculators/compressed-air-reduction']">Compressed Air Reduction</a>
-                <p>
-                    This calculator is used to quantify the energy savings associated with reducing compressed air
-                    usage.
-                </p>
-            </div>
-        </div>
-        <div class="d-flex calc-item-row w-100">
-            <div class="p-2 col-2 d-flex justify-content-center align-items-center">
-                <div class="calc-icon-div d-flex justify-content-center align-items-center">
                     <img src="assets/images/calculator-icons/utilities-icons/compressed-air-pressure-reduction-icon.png"
                         class="calc-img" [routerLink]="['/calculators/compressed-air-pressure-reduction']">
                 </div>
