+++ src/app/dashboard/sidebar/calculator-list/calculator-list.component.html
@@ -130,9 +130,6 @@
       class="sidebar-item">Electricity Reduction</span>
     <span [routerLink]="['/calculators/natural-gas-reduction']" routerLinkActive="selected" class="sidebar-item">Natural
       Gas Reduction</span>
-    <span [routerLink]="['/calculators/compressed-air-reduction']" routerLinkActive="selected"
-      class="sidebar-item">Compressed Air
-      Reduction</span>
     <span [routerLink]="['/calculators/compressed-air-pressure-reduction']" routerLinkActive="selected"
       class="sidebar-item">Compressed Air
       Pressure Reduction</span>
